conform the hardening to the patterns wiki
